Integrating web and mobile projects

+++ server/src/routes/memories.ts
@@ -3,8 +3,14 @@ import { z } from 'zod';
 import { prisma } from '../lib/prisma';
 
 export const memoriesRoutes = async (app: FastifyInstance) => {
-  app.get('/memories', async () => {
+  /* Antes de executar o preHandler(callback) de cada rota, verifique se o usuário esta autenticado.*/
+  app.addHook('preHandler', async request => await request.jwtVerify());
+
+  app.get('/memories', async request => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -18,18 +24,23 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     });
   });
 
-  app.get('/memories/:id', async request => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
 
-    /* Encontra a memoria com esse id, se não ele dispara um erro */
+    /* Encontra a memoria com esse id, caso não encontre ele dispara um erro */
     const memory = await prisma.memory.findFirstOrThrow({
       where: {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
     return memory;
   });
 
@@ -46,13 +57,13 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
         content,
         coverUrl,
         isPublic,
-        userId: 'cdea01a7-0a24-4384-82b5-7016b3d2194d',
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put('/memories/:id', async request => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -65,7 +76,17 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     });
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -78,16 +99,27 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     return memory;
   });
 
-  app.delete('/memories/:id', async request => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
+
     const { id } = paramsSchema.parse(request.params);
-    const memory = await prisma.memory.delete({
+
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     });
-    return memory;
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
+    });
   });
 };
